+++ src/code_8057824_1.c
@@ -246,8 +246,134 @@ extern u32 gUnknown_8105B68;
 extern u32 gUnknown_8105BA8;
 extern u8 *gUnknown_8105798[];
 extern u32 gJirachiReFightDialogue_1;
+extern u32 gUnknown_202EDC8;
+
+extern u32 gKyogrePreFightDialogue_8;
+extern u32 gKyogrePreFightDialogue_1;
+extern u32 gKyogrePreFightDialogue_2;
+extern u32 gKyogrePreFightDialogue_3;
+extern u32 gKyogrePreFightDialogue_4;
+extern u32 gKyogrePreFightDialogue_5;
+extern u32 gKyogrePreFightDialogue_6;
+extern u32 gKyogrePreFightDialogue_7;
+extern u32 gUnknown_810697C;
+extern u64 gUnknown_8107544[];
+extern u32 gMedichamRescueDialogue_1;
+extern u32 gMedichamRescueDialogue_2;
+extern u32 gMedichamRescueDialogue_3;
+extern u32 gMedichamRescueDialogue_4;
+extern u32 gSmeargleRescueDialogue_1;
+extern u32 gSmeargleRescueDialogue_2;
+extern u32 gSmeargleRescueDialogue_3;
+extern u32 gSmeargleRescueDialogue_4;
+extern u32 gUnknown_8106720;
+
+extern u32 gCelebiJoinDialogue_10;
+extern u32 gCelebiJoinDialogue_1;
+extern u32 gCelebiJoinDialogue_2;
+extern u32 gCelebiJoinDialogue_3;
+extern u32 gCelebiJoinDialogue_4;
+extern u32 gCelebiJoinDialogue_5;
+extern u32 gCelebiJoinDialogue_6;
+extern u32 gCelebiJoinDialogue_7;
+extern u32 gCelebiJoinDialogue_8;
+extern u32 gCelebiJoinDialogue_9;
+extern u32 *gPtrPurityForestAllowCelebiToJoinText;
+extern u32 gPurityForestAllowCelebiToJoinPrompt;
+extern u32 *gPtrPurityForestRefuseCelebiConfirmText;
+extern u32 gPurityForestRefuseCelebiConfirmPrompt;
+extern u32 gDeoxysPreFightDialogue_1;
+extern u32 gDeoxysPreFightDialogue_2;
+extern u32 gDeoxysPreFightDialogue_3;
+extern u32 gDeoxysPreFightDialogue_4;
+extern u32 gDeoxysPreFightDialogue_5;
+extern u32 gLugiaPreFightDialogue_8;
+extern u32 gLugiaPreFightDialogue_1;
+extern u32 gLugiaPreFightDialogue_2;
+extern u32 gLugiaPreFightDialogue_3;
+extern u32 gLugiaPreFightDialogue_4;
+extern u32 gLugiaPreFightDialogue_5;
+extern u32 gLugiaPreFightDialogue_6;
+extern u32 gLugiaPreFightDialogue_7;
+extern u32 gZapdosReFightDialogue_1;
+extern u32 gUnknown_81015A0;
+extern u32 gZapdosReFightDialogue_2;
+extern u32 gZapdosReFightDialogue_3;
+extern u32 gZapdosReFightDialogue_4;
+
+extern u32 gUnknown_81058C4;
+extern u32 gUnknown_81058A8;
 
 extern u8 gUnknown_81074FC[];
+
+extern u8 sub_800E9A8(u32);
+extern void sub_800DC14(u32);
+
+extern void sub_8086A3C(struct DungeonEntity *r0);
+extern void sub_8083E88(u32);
+extern void sub_80854D4(void);
+extern void sub_80855E4(void *);
+extern void sub_8085930(u32);
+extern void sub_8068FE0(struct DungeonEntity *, u32, u32);
+extern void sub_8085860(s32 r0, u32 r1);
+extern void sub_8083EA8(u32, u32);
+
+extern void sub_803E708(u32, u32);
+extern void sub_8086448(void);
+extern void sub_8086598(void);
+extern void SpriteLookAroundEffect(struct DungeonEntity *);
+extern void SpriteShockEffect(struct DungeonEntity *);
+extern void PlaySoundEffect(u32);
+extern void sub_806CDD4(struct DungeonEntity *, u32, u32);
+extern void sub_80869E4(struct DungeonEntity *, u32, u32, u32);
+extern void sub_806CE68(struct DungeonEntity *, u32);
+extern void sub_804539C(struct DungeonEntity *, u32, u32);
+extern void sub_8042B0C(struct DungeonEntity *);
+extern void SetFacingDirection(struct DungeonEntity *, u32);
+extern void DisplayDungeonDialogue(u32 *);
+extern void sub_803E708(u32, u32);
+extern u8 HasRecruitedMon(u32);
+extern u8 sub_806FD18(struct DungeonEntity *);
+extern u8 sub_8083E74(u32);
+extern void sub_8083F14();
+extern void sub_80861D4(struct DungeonEntity *, u32, u32);
+extern void sub_806FDF4(struct DungeonEntity *, struct DungeonEntity *, struct DungeonEntity **);
+extern u32 sub_80861F8(u32, struct DungeonEntity *, u32);
+extern void sub_8083ED8(u32);
+extern s32 GetCameraXPos();
+extern s32 GetCameraYPos();
+extern void sub_803F878(u32, s32);
+extern void SetupBossFightHP(struct DungeonEntity *, u32, u32);
+extern void sub_803E9D0(void);
+extern void DeoxysScreenFlash(void);
+extern void sub_8085EB0();
+extern void sub_808563C(void *);
+
+void SetupDeoxysFightHP(struct DungeonEntity *r0);
+void sub_808C550(void);
+void sub_808C590(struct DungeonEntity *r0);
+void SetupDeoxysFacingDirection(struct DungeonEntity *r0);
+void sub_808C9B0();
+void sub_808C8E0(struct DungeonEntity *param_1);
+void sub_808C360(void);
+void KyogreScreenFlash();
+
+void sub_808C0CC();
+void LugiaScreenFlash();
+void LugiaScreenFlash2();
+extern void sub_80861B8(struct DungeonEntity *, u32, u32);
+
+void nullsub_99();
+extern bool8 *gFriendAreas;
+
+extern void sub_8049884();
+extern void sub_8049B8C();
+extern void sub_8049ED4();
+extern void sub_8040A84();
+extern void sub_8086A54(struct DungeonEntity *);
+
+extern void sub_806BFC0(struct DungeonEntityData *, u32);
+
 
 extern struct DungeonEntity *xxx_call_GetLeaderEntity(void);
 extern struct DungeonEntity *GetEntityFromClientType(u32);
@@ -259,28 +385,21 @@ extern void sub_8083ED8(u32);
 extern void sub_803E708(u32, u32);
 extern void DisplayDungeonDialogue(u32 *r0);
 extern void sub_8083E88(u32);
-extern void JirachiSpinEffect(struct DungeonEntity *);
-extern void JirachiWishGrantDialogue(struct DungeonEntity *jirachiEntity);
 
-extern void sub_80862BC(struct DungeonEntity *);
-extern void sub_8086448();
-extern void SpriteLookAroundEffect(struct DungeonEntity *);
+extern void SpriteShockEffect(struct DungeonEntity *);
 extern u32 sub_80861F8(u32, struct DungeonEntity *, u32);
-extern void SetupBossFightHP(struct DungeonEntity *, u32, u32);
 
 extern void sub_8097FA8(u32);
+
 extern void JirachiWish();
+extern void JirachiSpinEffect(struct DungeonEntity *);
+extern void JirachiWishGrantDialogue(struct DungeonEntity *jirachiEntity);
+extern void JirachiWishGrantFlash();
+extern void JirachiDropInEffect(struct DungeonEntity *param_1);
+extern u8 JirachiFriendAreaSearch(void);
+
 extern u8 sub_80860A8(u32);
-extern void sub_8040A84();
-extern void sub_8086A3C(struct DungeonEntity *);
-extern u8 HasRecruitedMon(u32);
-extern void sub_80854D4();
-extern void sub_8085930(u32);
-extern void sub_8068FE0(struct DungeonEntity *, u32, u32);
-extern void SetFacingDirection(struct DungeonEntity *, u32);
 extern void sub_8049ED4();
-extern void sub_8085860(s32 r0, u32 r1);
-extern void sub_806CE68(struct DungeonEntity *, u32);
 extern void sub_8092578(u8 *buffer, u8 index, u8 r2);
 extern void sub_8052D44(u8 *, struct DungeonEntity *, struct DungeonEntity *);
 extern void sub_8097FD0(u32);
@@ -291,10 +410,8 @@ extern void sub_8046D20();
 extern void sub_808B1CC(u8);
 extern void sub_80464C8(struct DungeonEntity *, u32 *, struct ItemSlot *);
 extern void sub_8098044(u32);
-extern void PlaySoundEffect(u32);
 extern void SetDungeonBGColorRGB(u32, u32, u32, u32, u32);
 extern void sub_803E46C(u32);
-extern void sub_8085EB0();
 extern void sub_80866C4(u32 *);
 extern void sub_808692C();
 
@@ -310,9 +427,7 @@ extern void sub_8086838(u32, u32, u32);
 extern void sub_80426C8(u32, u32);
 extern void sub_807EAA0(u32, u32);
 
-extern void sub_808563C(void *);
 extern void sub_808BCE4(void);
-extern void JirachiDropInEffect(struct DungeonEntity *param_1);
 extern u8 sub_8098100(u32);
 extern void sub_8086794();
 
@@ -320,7 +435,7 @@ void EnteiScreenFlash(void);
 void RaikouScreenFlash(void);
 void sub_808A528(struct DungeonEntity * param_1);
 void SuicuneScreenFlash();
-void sub_808A904(struct DungeonEntity * param_1);
+void HoOhDropInEffect(struct DungeonEntity * param_1);
 void HoOhScreenFlash(void);
 void LatiosScreenFlash(void);
 void SetupRegiFacingDirection(struct DungeonEntity *r0);
@@ -328,18 +443,14 @@ void SetupRegirockFightHP(struct DungeonEntity *r0);
 void SetupRegiceFightHP(struct DungeonEntity *r0);
 void SetupRegisteelFightHP(struct DungeonEntity *r0);
 void sub_808B50C(void);
-void sub_8089B64(struct DungeonEntity *param_1);
+void MewtwoDropInEffect(struct DungeonEntity *param_1);
 void MewtwoScreenFlash(void);
 void sub_80898F8(struct DungeonEntity *r0);
 void sub_8089908(struct DungeonEntity *r0);
-void sub_8089620(struct DungeonEntity *param_1);
+void RayquazaDropInEffect(struct DungeonEntity *param_1);
 void RayquazaScreenFlash(void);
 void sub_80891F0(void);
 void sub_8089294(void);
-void sub_80891B0();
-void sub_8089224();
-void sub_80891D0();
-void sub_808919C();
 void GroudonScreenFlash2(void);
 
 extern void sub_8083F14();
@@ -385,19 +496,13 @@ extern void sub_8086500(void);
 extern void sub_8086690(void);
 void MoltresScreenDarken(void);
 extern void sub_808663C(void);
-void sub_8087704(struct DungeonEntity *param_1);
+void ZapdosDropInEffect(struct DungeonEntity *param_1);
 void ZapdosScreenFlash(int param_1);
 
-extern u32 gZapdosReFightDialogue_1;
-extern u32 gUnknown_81015A0;
-extern u32 gZapdosReFightDialogue_2;
-extern u32 gZapdosReFightDialogue_3;
-extern u32 gZapdosReFightDialogue_4;
 extern s32 sub_8052C68(u32, u8 *, u32 *, u32);
 extern void sub_80861A8(void);
 extern void sub_8045C28(struct ItemSlot *, u8 , u8 *);
 extern void sub_808BD38(void);
-extern u8 JirachiFriendAreaSearch(void);
 extern void sub_808BB3C(s16 *);
 extern void sub_8046860(struct DungeonEntity *, s32 *, struct ItemSlot *, u32);
 extern u32 sub_803D73C(u32);
@@ -416,7 +521,7 @@ void ZapdosReFightDialogue(void)
   PlaySoundEffect(0x1da);
   DisplayDungeonDialogue(&gUnknown_81015A0);
   sub_803E708(10,0x46);
-  sub_8087704(ZapdosEntity);
+  ZapdosDropInEffect(ZapdosEntity);
   sub_803E708(10,0x46);
   DisplayDungeonDialogue(&gZapdosReFightDialogue_2);
   ZapdosScreenFlash(1);
@@ -442,7 +547,7 @@ void ZapdosPostStoryPreFightDialogue(void)
     ZapdosScreenFlash(1);
     PlaySoundEffect(0x1da);
     DisplayDungeonDialogue(&gZapdosPostStoryPreFightDialogue_1);
-    sub_8087704(ZapdosEntity);
+    ZapdosDropInEffect(ZapdosEntity);
     DisplayDungeonDialogue(&gZapdosPostStoryPreFightDialogue_2);
     ZapdosScreenFlash(1);
     DisplayDungeonDialogue(&gZapdosPostStoryPreFightDialogue_3);
@@ -457,17 +562,17 @@ void ZapdosPostStoryPreFightDialogue(void)
 }
 
 
-void sub_8087704(struct DungeonEntity *param_1)
+void ZapdosDropInEffect(struct DungeonEntity *zapdosEntity)
 {
   int iVar1;
 
-  param_1->entityData->unk15C = 1;
-  param_1->entityData->unk15E = 0;
-  param_1->entityData->unk174 = 200;
+  zapdosEntity->entityData->unk15C = 1;
+  zapdosEntity->entityData->unk15E = 0;
+  zapdosEntity->entityData->unk174 = 200;
   PlaySoundEffect(0x1ea);
   for(iVar1 = 200; iVar1 >= 0; iVar1 -= 5)
   {
-    param_1->entityData->unk174 = iVar1 * 256;
+    zapdosEntity->entityData->unk174 = iVar1 * 256;
     sub_803E46C(0x46);
   }
   sub_803E708(0x1e,0x46);
@@ -715,17 +820,17 @@ void MoltresPostStoryPreFightDialogue(void)
   }
 }
 
-void MoltresDropInEffect(struct DungeonEntity * param_1)
+void MoltresDropInEffect(struct DungeonEntity * moltresEntity)
 {
   s32 iVar1;
 
-  param_1->entityData->unk15C = 1;
-  param_1->entityData->unk15E = 0;
-  param_1->entityData->unk174 = 0xc800;
+  moltresEntity->entityData->unk15C = 1;
+  moltresEntity->entityData->unk15E = 0;
+  moltresEntity->entityData->unk174 = 0xc800;
   PlaySoundEffect(0x1f8);
   for(iVar1 = 200; iVar1 >= 0; iVar1 -= 5)
   {
-    param_1->entityData->unk174 = iVar1 * 256;
+    moltresEntity->entityData->unk174 = iVar1 * 256;
     sub_803E46C(0x46);
   }
 }
@@ -1872,7 +1977,7 @@ void RayquazaPreFightDialogue(void)
   RayquazaScreenFlash();
   PlaySoundEffect(0x1da); // Rayquaza Cry
   DisplayDungeonDialogue(&gRayquazaPreFightDialogue_7);
-  sub_8089620(RayquazaEntity);
+  RayquazaDropInEffect(RayquazaEntity);
   DisplayDungeonDialogue(&gRayquazaPreFightDialogue_8);
   RayquazaScreenFlash();
   DisplayDungeonDialogue(&gRayquazaPreFightDialogue_9);
@@ -1896,7 +2001,7 @@ void RayquazaReFightDialogue(void)
   PlaySoundEffect(0x1da); // Rayquaza Cry
   DisplayDungeonDialogue(&gRayquazaPreFightDialogue_7);
   sub_803E708(10,0x46);
-  sub_8089620(RayquazaEntity);
+  RayquazaDropInEffect(RayquazaEntity);
   DisplayDungeonDialogue(&gRayquazaReFightDialogue_1);
   RayquazaScreenFlash();
   DisplayDungeonDialogue(&gRayquazaReFightDialogue_2);
@@ -1922,7 +2027,7 @@ void RayquazaPostStoryPreFightDialogue(void)
   else {
     PlaySoundEffect(0x1da); // Rayquaza Cry
     DisplayDungeonDialogue(&gRayquazaPostStoryPreFightDialogue_1);
-    sub_8089620(RayquazaEntity);
+    RayquazaDropInEffect(RayquazaEntity);
     DisplayDungeonDialogue(&gRayquazaPostStoryPreFightDialogue_2);
     RayquazaScreenFlash();
     DisplayDungeonDialogue(&gRayquazaPostStoryPreFightDialogue_3);
@@ -1935,12 +2040,12 @@ void RayquazaPostStoryPreFightDialogue(void)
   }
 }
 
-void sub_8089620(struct DungeonEntity *param_1)
+void RayquazaDropInEffect(struct DungeonEntity *rayquazaEntity)
 {
   s32 iVar1;
   s32 iVar2;
 
-  param_1->entityData->unk15E = 0;
+  rayquazaEntity->entityData->unk15E = 0;
   iVar2 = 51200;
   iVar1 = 0x600;
   PlaySoundEffect(0x1f8);
@@ -1951,10 +2056,10 @@ void sub_8089620(struct DungeonEntity *param_1)
       iVar1 = 0x14;
     }
     if (iVar2 < 0) break;
-    param_1->entityData->unk174 = iVar2;
+    rayquazaEntity->entityData->unk174 = iVar2;
     sub_803E46C(0x46);
   }
-  param_1->entityData->unk174 = 0;
+  rayquazaEntity->entityData->unk174 = 0;
 }
 
 void RayquazaScreenFlash(void)
@@ -2040,7 +2145,7 @@ void sub_80897F0(void)
   LeaderEntity = xxx_call_GetLeaderEntity();
   sub_8086448();
   sub_803E708(10,0x46);
-  sub_808563C(sub_80862BC);
+  sub_808563C(SpriteShockEffect);
   sub_803E708(0x20,0x46);
   sub_808692C();
   DisplayDungeonDialogue(&gUnknown_8103BD8);
@@ -2077,7 +2182,7 @@ void sub_8089878(void)
   LeaderEntity = xxx_call_GetLeaderEntity();
   sub_8086448();
   sub_803E708(10,0x46);
-  sub_808563C(sub_80862BC);
+  sub_808563C(SpriteShockEffect);
   sub_803E708(0x20,0x46);
   sub_808692C();
   DisplayDungeonDialogue(&gUnknown_8103D8C);
@@ -2167,7 +2272,7 @@ void MewtwoPreFightDialogue(void)
   sub_8083E88(0x23);
   sub_8086794();
   sub_803E708(0x1e,0x46);
-  sub_8089B64(MewtwoEntity);
+  MewtwoDropInEffect(MewtwoEntity);
   sub_803E708(0x3c,0x46);
   DisplayDungeonDialogue(&gMewtwoPreFightDialogue_2);
   DisplayDungeonDialogue(&gMewtwoPreFightDialogue_3);
@@ -2196,7 +2301,7 @@ void MewtwoReFightDialogue(void)
     DisplayDungeonDialogue(&gMewtwoReFightDialogue_1);
     sub_8083E88(0x23);
     sub_8086794();
-    sub_8089B64(MewtwoEntity);
+    MewtwoDropInEffect(MewtwoEntity);
     sub_803E708(0x3c,0x46);
     DisplayDungeonDialogue(&gMewtwoReFightDialogue_2);
     MewtwoScreenFlash();
@@ -2208,26 +2313,26 @@ void MewtwoReFightDialogue(void)
   }
 }
 
-void sub_8089B64(struct DungeonEntity *param_1)
+void MewtwoDropInEffect(struct DungeonEntity *mewtwoEntity)
 {
   s32 iVar1;
   s32 iVar2;
 
-  param_1->entityData->unk15E = 0;
+  mewtwoEntity->entityData->unk15E = 0;
   iVar2 = 51200;
   iVar1 = 0x400;
   PlaySoundEffect(0x1f8);
   while( 1 ) {
     iVar2 = iVar2 - iVar1;
-    iVar1 = iVar1 -= 11;
+    iVar1 -= 11;
     if (iVar1 < 0x1e) {
       iVar1 = 0x1e;
     }
     if (iVar2 < 0) break;
-    param_1->entityData->unk174 = iVar2;
+    mewtwoEntity->entityData->unk174 = iVar2;
     sub_803E46C(0x46);
   }
-  param_1->entityData->unk174 = 0;
+  mewtwoEntity->entityData->unk174 = 0;
 }
 
 void MewtwoScreenFlash(void)
@@ -2794,7 +2899,7 @@ void HoOhPreFightDialogue(void)
   sub_808680C();
   PlaySoundEffect(0x1c5);
   sub_803E708(0xfa,70);
-  sub_808A904(HoOhEntity);
+  HoOhDropInEffect(HoOhEntity);
   sub_8083E58(0x1c5);
   sub_808682C();
   sub_80855E4(sub_8086910);
@@ -2834,7 +2939,7 @@ void HoOhReFightDialogue(void)
     sub_80855E4(sub_80868F4);
     sub_808680C();
     sub_803E708(10,70);
-    sub_808A904(HoOhEntity);
+    HoOhDropInEffect(HoOhEntity);
     sub_808682C();
     sub_80855E4(sub_8086910);
     sub_8085930(DIRECTION_NORTH);
@@ -2849,7 +2954,7 @@ void HoOhReFightDialogue(void)
   }
 }
 
-void sub_808A904(struct DungeonEntity * param_1)
+void HoOhDropInEffect(struct DungeonEntity * param_1)
 {
   s32 iVar1;
   s32 iVar2;
@@ -2972,7 +3077,7 @@ void LatiosPreFightDialogue(void)
   sub_8086448();
   ShiftCameraToPosition(&local_18,0x18);
   sub_803E708(2,70);
-  sub_80862BC(LatiosEntity);
+  SpriteShockEffect(LatiosEntity);
   sub_803E708(0x20,70);
   sub_808692C();
   DisplayDungeonDialogue(&gLatiosPreFightDialogue_1);
@@ -3383,10 +3488,10 @@ void sub_808B1CC(u8 r0)
 }
 
 
-void SetupRegiFacingDirection(struct DungeonEntity *r0)
+void SetupRegiFacingDirection(struct DungeonEntity *regiEntity)
 {
-    r0->entityData->action.facingDir = DIRECTION_NORTH;
-    sub_806CE68(r0, DIRECTION_NORTH);
+    regiEntity->entityData->action.facingDir = DIRECTION_NORTH;
+    sub_806CE68(regiEntity, DIRECTION_NORTH);
 }
 
 
@@ -3555,7 +3660,7 @@ void JirachiPreFightDialogue(void)
   sub_803E708(10,70);
   DisplayDungeonDialogue(&gJirachiPreFightDialogue_2);
   sub_80861F8(0x37,JirachiEntity,1);
-  sub_80855E4(sub_80862BC);
+  sub_80855E4(SpriteShockEffect);
   sub_803E708(0x3c,70);
   JirachiDropInEffect(JirachiEntity);
   sub_803E708(0x5a,70);
@@ -3631,7 +3736,7 @@ void JirachiWish(void)
   sub_803E708(10,0x46);
   PlaySoundEffect(0x1a7);
   sub_803E708(0x96,0x46);
-  sub_80862BC(JirachiEntity);
+  SpriteShockEffect(JirachiEntity);
   DisplayDungeonDialogue(&gUnknown_8105674);
   sub_8083E88(0x79);
   sub_803E708(10,0x46);
@@ -3870,4 +3975,1169 @@ void sub_808BB3C(s16 *param_1)
 
   local_8 = (iVar1) | (uVar2 & 0x0000ffff);
   sub_8085EC8(100,0,0,&local_8,0);
+}
+
+void sub_808BBA8(struct DungeonEntity *jirachiEntity)
+{
+  sub_806BFC0(jirachiEntity->entityData,0);
+}
+
+void JirachiDropInEffect(struct DungeonEntity *jirachiEntity)
+{
+  s32 iVar1;
+  s32 iVar2;
+
+  sub_80861F8(0x1b,jirachiEntity,0);
+  sub_8086A54(jirachiEntity);
+  sub_80861B8(jirachiEntity,0xe,0);
+  iVar1 = 0xa000;
+  iVar2 = 0x200;
+  PlaySoundEffect(0x1f8);
+  while( 1 ) {
+    iVar1 = iVar1 - iVar2;
+    if (iVar1 < 0x1800) {
+      iVar2 = 0x100;
+    }
+    if (iVar1 < 0) break;
+    jirachiEntity->entityData->unk174 = iVar1;
+    sub_803E46C(0x46);
+  }
+ jirachiEntity->entityData->unk174 = 0;
+}
+
+void JirachiSpinEffect(struct DungeonEntity * jirachiEntity)
+{
+  s32 uVar1;
+
+  PlaySoundEffect(0x298);
+  for(uVar1 = 0; uVar1 < 25; uVar1++){
+    SetFacingDirection(jirachiEntity, uVar1 & DIRECTION_MASK);
+    sub_803E708(3,0x46);
+  }
+  jirachiEntity->entityData->unk15E = 1;
+  PlaySoundEffect(0x27f);
+  sub_80861F8(99,jirachiEntity,1);
+}
+
+void JirachiWishGrantDialogue(struct DungeonEntity *jirachiEntity)
+{
+  u32 uVar2;
+
+  jirachiEntity->entityData->unk15D = 1;
+  // Nnnnnnnnnn!
+  DisplayDungeonDialogue(&gUnknown_81058A8);
+  PlaySoundEffect(0x375);
+  sub_80861F8(0x67,jirachiEntity,1);
+  PlaySoundEffect(0x2a8);
+  uVar2 = sub_80861F8(0x68,jirachiEntity,0);
+  // Taaaaaaaah!
+  DisplayDungeonDialogue(&gUnknown_81058C4);
+  if (sub_800E9A8(uVar2) != 0) {
+    sub_800DC14(uVar2);
+  }
+  sub_803E708(10,0x46);
+  JirachiWishGrantFlash();
+}
+
+void sub_808BCE4(void)
+{
+  struct MapTile *puVar1;
+
+  puVar1 = GetMapTile_2(gDungeonGlobalData->unkE23C, gDungeonGlobalData->unkE23E);
+  puVar1->tileType &= ~(TILE_TYPE_FLOOR | TILE_TYPE_LIQUID);
+  puVar1->tileType |= TILE_TYPE_MAP_EDGE;
+  puVar1->tileType &= ~TILE_TYPE_STAIRS;
+  sub_8049884();
+  sub_8049B8C();
+  sub_8049ED4();
+  sub_8040A84();
+}
+
+void sub_808BD38(void)
+{
+  struct MapTile *puVar1;
+
+  puVar1 = GetMapTile_2(gDungeonGlobalData->unkE23C, gDungeonGlobalData->unkE23E);
+  puVar1->tileType &= ~(TILE_TYPE_FLOOR | TILE_TYPE_LIQUID);
+  puVar1->tileType |= TILE_TYPE_FLOOR;
+  puVar1->tileType &= ~TILE_TYPE_MAP_EDGE;
+  puVar1->tileType |= TILE_TYPE_STAIRS;
+  puVar1->unk8 = 1;
+  sub_8049884();
+  sub_8049B8C();
+  sub_8049ED4();
+  sub_8040A84();
+}
+
+u8 JirachiFriendAreaSearch(void)
+{
+  u8 unlockCondition;
+  s32 friendAreaCounter;
+  s32 numUnlockableAreas;
+  u8 friendAreas[NUM_FRIEND_AREAS];
+  u32 friendAreaIndex;
+
+  // Build a copy of friend area list and keep track of
+  // how many are only Story Unlockable
+  numUnlockableAreas = 0;
+  for(friendAreaCounter = 1; friendAreaCounter < NUM_FRIEND_AREAS; friendAreaCounter++){
+    unlockCondition = GetFriendAreaUnlockCondition(friendAreaCounter);
+    if (!(unlockCondition == UNLOCK_LEGENDARY_REQUEST)){
+        if (!(unlockCondition == UNLOCK_WONDER_MAIL)){
+            if(!gFriendAreas[friendAreaCounter]) {
+                friendAreas[numUnlockableAreas] = friendAreaCounter;
+                numUnlockableAreas++;
+            }
+        }
+    }
+  }
+
+  if (numUnlockableAreas == 0) {
+    return NUM_FRIEND_AREAS;
+  }
+  else {
+      // Randomly pick 1
+    friendAreaIndex = DungeonRandomCapped(numUnlockableAreas);
+    return friendAreas[friendAreaIndex];
+  }
+}
+
+void sub_808BDEC(void)
+{
+  struct DungeonEntity * LeaderEntity;
+  struct DungeonEntity * LugiaEntity;
+
+  LeaderEntity = xxx_call_GetLeaderEntity();
+  LugiaEntity = GetEntityFromClientType(0x1b);
+  sub_8083F14();
+  sub_80854D4();
+  sub_8085930(DIRECTION_NORTH);
+  sub_80855E4(sub_8086A3C);
+  if (HasRecruitedMon(SPECIES_LUGIA)) {
+    sub_8083E88(0x23);
+    sub_8068FE0(LugiaEntity,0x21c,0);
+  }
+  else {
+    gDungeonGlobalData->unk7 = 1;
+    SetFacingDirection(LugiaEntity, DIRECTION_SOUTH);
+  }
+  sub_8085860(LeaderEntity->posWorld.x,LeaderEntity->posWorld.y - 3);
+  CopySpeciesNametoBuffer(gUnknown_202E038, SPECIES_LUGIA);
+}
+
+void sub_808BE70(u8 param_1,u8 param_2)
+{
+  if ((param_2 == 0x33) && (param_1 == 0x1B)) {
+    gDungeonGlobalData->unk2 = 1;
+  }
+}
+
+
+void LugiaPreFightDialogue(void)
+{
+  struct DungeonEntity * LeaderEntity;
+  struct DungeonEntity * LugiaEntity;
+
+  LeaderEntity = xxx_call_GetLeaderEntity();
+  LugiaEntity = GetEntityFromClientType(0x1b);
+  if (HasRecruitedMon(SPECIES_LUGIA)) {
+    sub_8086448();
+    sub_80866C4(&gLugiaPreFightDialogue_8);
+  }
+  else {
+    sub_8083F14();
+    sub_8086448();
+    DisplayDungeonDialogue(&gLugiaPreFightDialogue_1);
+    sub_8083EA8(0x23,0x3c);
+    sub_808C0CC();
+    DisplayDungeonDialogue(&gLugiaPreFightDialogue_2);
+    LugiaScreenFlash();
+    DisplayDungeonDialogue(&gLugiaPreFightDialogue_3);
+    sub_803E708(10,0x46);
+    DisplayDungeonDialogue(&gLugiaPreFightDialogue_4);
+    PlaySoundEffect(0x1f8);
+    sub_80861D4(LugiaEntity,0xd,0);
+    sub_803E708(0x2b,0x46);
+    sub_80861B8(LugiaEntity,0,0);
+    DisplayDungeonDialogue(&gLugiaPreFightDialogue_5);
+    LugiaScreenFlash();
+    DisplayDungeonDialogue(&gLugiaPreFightDialogue_6);
+    LugiaScreenFlash2();
+    SetDungeonBGColorRGB(0xffffff06,0xffffff06,0xffffff06,1,0);
+    sub_8083F14();
+    sub_80861D4(LugiaEntity,7,0);
+    DisplayDungeonDialogue(&gLugiaPreFightDialogue_7);
+    LugiaScreenFlash();
+    SetupBossFightHP(LugiaEntity,800,0x20);
+    sub_8083E88(0x23);
+    ShiftCameraToPosition(&LeaderEntity->posPixel,0x10);
+  }
+}
+
+void LugiaScreenFlash(void)
+{
+  s32 iVar1;
+
+  PlaySoundEffect(0x1f6);
+  for(iVar1 = 250; iVar1 > 149; iVar1 -= 10){
+    SetDungeonBGColorRGB(0,0,iVar1,1,1);
+    sub_803E46C(0x46);
+  }
+  sub_803E708(10,0x46);
+  for(iVar1 = 250; iVar1 > 199; iVar1 -= 10){
+    SetDungeonBGColorRGB(0,iVar1,iVar1,1,1);
+    sub_803E46C(0x46);
+  }
+  sub_803E708(10,0x46);
+  for(iVar1 = 250; iVar1 >= 0; iVar1 -= 10){
+    SetDungeonBGColorRGB(0,0,iVar1,1,1);
+    sub_803E46C(0x46);
+  }
+  sub_803E708(10,0x46);
+  sub_8085EB0();
+}
+
+void LugiaScreenFlash2(void)
+{
+  s32 iVar1;
+
+  PlaySoundEffect(0x1f6);
+
+  for(iVar1 = 0; iVar1 < 200; iVar1 += 100){
+    SetDungeonBGColorRGB(iVar1,iVar1,iVar1,1,0);
+    sub_803E46C(0x46);
+  }
+
+  for(iVar1 = 200; iVar1 >= 0; iVar1 -= 100){
+    SetDungeonBGColorRGB(iVar1,iVar1,iVar1,1,0);
+    sub_803E46C(0x46);
+  }
+
+  sub_803E708(4,0x46);
+
+  for(iVar1 = 0; iVar1 < 200; iVar1 += 100){
+    SetDungeonBGColorRGB(iVar1,iVar1,iVar1,1,0);
+    sub_803E46C(0x46);
+  }
+
+  for(iVar1 = 200; iVar1 >= 0; iVar1 -= 100){
+    SetDungeonBGColorRGB(iVar1,iVar1,iVar1,1,0);
+    sub_803E46C(0x46);
+  }
+
+  sub_803E708(10,0x46);
+  sub_8085EB0();
+}
+
+void sub_808C0CC(void)
+{
+  gUnknown_202EDC8 = 0x1f;
+  SetDungeonBGColorRGB(0,0,0,1,1);
+  sub_803E9D0();
+  sub_803E46C(0x46);
+  gDungeonGlobalData->unk7 = 0;
+  LugiaScreenFlash();
+}
+
+void sub_808C10C(void)
+{
+  struct DungeonEntity * LeaderEntity;
+  struct DungeonEntity * KyogreEntity;
+
+  u32 uVar3;
+  s32 iVar2;
+
+  LeaderEntity = xxx_call_GetLeaderEntity();
+  KyogreEntity = GetEntityFromClientType(0x1c);
+  sub_8083F14();
+  sub_80854D4();
+  sub_8085930(DIRECTION_NORTH);
+  if (HasRecruitedMon(SPECIES_KYOGRE)) {
+    sub_8068FE0(KyogreEntity,0x21c,0);
+  }
+  else {
+    gDungeonGlobalData->unk7 = 1;
+    SetFacingDirection(KyogreEntity, DIRECTION_SOUTH);
+  }
+  sub_8085860(LeaderEntity->posWorld.x,LeaderEntity->posWorld.y);
+  uVar3 = GetCameraXPos();
+  iVar2 = GetCameraYPos();
+  sub_803F878(uVar3,iVar2 + -0x1000);
+  CopySpeciesNametoBuffer(gUnknown_202E038,SPECIES_KYOGRE);
+  // TODO: gUnknown_202E088 doesn't match but gUnknown_202E038 + 0x50 does
+  CopySpeciesNametoBuffer(gUnknown_202E038 + 0x50, SPECIES_GROUDON);
+}
+
+void sub_808C1A4(u8 param_1,u8 param_2)
+{
+  if ((param_2 == 0x34) && (param_1 == 0x1C)) {
+    gDungeonGlobalData->unk2 = 1;
+    sub_8083E88(0x23);
+  }
+}
+
+void KyogrePreFightDialogue(void)
+{
+  struct DungeonEntity *LeaderEntity;
+  struct DungeonEntity *KyogreEntity;
+
+  LeaderEntity = xxx_call_GetLeaderEntity();
+  KyogreEntity = GetEntityFromClientType(0x1c);
+  sub_8083F14();
+  if (HasRecruitedMon(SPECIES_KYOGRE)) {
+    sub_8083EA8(0x23,0x3c);
+    SpriteLookAroundEffect(LeaderEntity);
+    sub_803E708(10,0x46);
+    DisplayDungeonDialogue(&gKyogrePreFightDialogue_8);
+    sub_803E708(10,0x46);
+    gDungeonGlobalData->unk2 = 1;
+  }
+  else {
+    DisplayDungeonDialogue(&gKyogrePreFightDialogue_1);
+    sub_803E708(10,0x46);
+    DisplayDungeonDialogue(&gKyogrePreFightDialogue_2);
+    sub_803E708(10,0x46);
+    nullsub_99();
+    sub_803E708(10,0x46);
+    // Gwwwwwooooooooh
+    DisplayDungeonDialogue(&gKyogrePreFightDialogue_3);
+    sub_803E708(10,0x46);
+    sub_8083EA8(0x23,0x3c);
+    sub_808C360();
+    // My duels against {ARG_POKEMON_3} (Groudon) left us both exhausted
+    DisplayDungeonDialogue(&gKyogrePreFightDialogue_4);
+    KyogreScreenFlash();
+    // I am {ARG_POKEMON_2} (Kyogre)!
+    // The lord of the sea!
+    DisplayDungeonDialogue(&gKyogrePreFightDialogue_5);
+    KyogreScreenFlash();
+    // Witness the destructive force of my waves!
+    DisplayDungeonDialogue(&gKyogrePreFightDialogue_6);
+    KyogreScreenFlash();
+    // Marvel at my awesome
+    // power!
+    DisplayDungeonDialogue(&gKyogrePreFightDialogue_7);
+    sub_803E708(10,0x46);
+    SetupBossFightHP(KyogreEntity,600,0xb);
+    sub_8083E88(0x23);
+    ShiftCameraToPosition(&LeaderEntity->posPixel,0x10);
+  }
+}
+
+void nullsub_99(void)
+{
+}
+
+void KyogreScreenFlash(void)
+{
+  s32 iVar1;
+
+  PlaySoundEffect(0x1f8);
+  for(iVar1 = 250; iVar1 > 149; iVar1 -= 10){
+    SetDungeonBGColorRGB(0,0,iVar1,1,1);
+    sub_803E46C(0x46);
+  }
+  sub_803E708(10,0x46);
+  for(iVar1 = 250; iVar1 > 199; iVar1 -= 10){
+    SetDungeonBGColorRGB(0,iVar1,iVar1,1,1);
+    sub_803E46C(0x46);
+  }
+  sub_803E708(10,0x46);
+  for(iVar1 = 250; iVar1 >= 0; iVar1 -= 10){
+    SetDungeonBGColorRGB(0,0,iVar1,1,1);
+    sub_803E46C(0x46);
+  }
+  sub_803E708(10,0x46);
+  sub_8085EB0();
+}
+
+void sub_808C360(void)
+{
+  gUnknown_202EDC8 = 0x1f;
+  SetDungeonBGColorRGB(0,0,0,1,1);
+  sub_803E9D0();
+  sub_803E46C(0x46);
+  gDungeonGlobalData->unk7 = 0;
+  KyogreScreenFlash();
+}
+
+void sub_808C3A0(void)
+{
+  struct DungeonEntity * LeaderEntity;
+
+  LeaderEntity = xxx_call_GetLeaderEntity();
+  sub_8083F14();
+  sub_80854D4();
+  sub_8085930(DIRECTION_NORTH);
+  sub_80855E4(sub_8086A3C);
+  if (HasRecruitedMon(SPECIES_DEOXYS_NORMAL)) {
+    sub_808563C(sub_808C590);
+  }
+  else {
+    gDungeonGlobalData->unk7 = 1;
+    sub_808563C(SetupDeoxysFacingDirection);
+  }
+  sub_8085860(LeaderEntity->posWorld.x,LeaderEntity->posWorld.y - 3);
+  CopySpeciesNametoBuffer(gUnknown_202E038, SPECIES_DEOXYS_NORMAL);
+}
+
+void sub_808C414(u8 param_1,u8 param_2)
+{
+  if ((param_2 == 0x35) && (param_1 == 0x1D)) {
+    gDungeonGlobalData->unk2 = 1;
+    sub_8083E88(0x1a);
+  }
+}
+
+void DeoxysPreFightDialogue(void)
+{
+  struct DungeonEntity * LeaderEntity;
+
+  LeaderEntity = xxx_call_GetLeaderEntity();
+  sub_8086448();
+  if (HasRecruitedMon(SPECIES_DEOXYS_NORMAL)) {
+    // There appears to be no one here.
+    // It's impossible to go any further
+    sub_80866C4(&gDeoxysPreFightDialogue_5);
+  }
+  else {
+    DisplayDungeonDialogue(&gDeoxysPreFightDialogue_1);
+    sub_8083EA8(0x1a,0x3c);
+    sub_808C550();
+    DisplayDungeonDialogue(&gDeoxysPreFightDialogue_2);
+    DeoxysScreenFlash();
+    DisplayDungeonDialogue(&gDeoxysPreFightDialogue_3);
+    DeoxysScreenFlash();
+    DisplayDungeonDialogue(&gDeoxysPreFightDialogue_4);
+    sub_803E708(10,0x46);
+    sub_808563C(SetupDeoxysFightHP);
+    ShiftCameraToPosition(&LeaderEntity->posPixel,0x10);
+  }
+}
+
+void DeoxysScreenFlash(void)
+{
+  s32 iVar1;
+
+  PlaySoundEffect(0x2c1);
+  for(iVar1 = 250; iVar1 > 149; iVar1 -= 10){
+    SetDungeonBGColorRGB(iVar1,0,0,1,1);
+    sub_803E46C(0x46);
+  }
+  sub_803E708(10,0x46);
+  for(iVar1 = 250; iVar1 > 199; iVar1 -= 10){
+    SetDungeonBGColorRGB(iVar1,iVar1,0,1,1);
+    sub_803E46C(0x46);
+  }
+  sub_803E708(10,0x46);
+  for(iVar1 = 250; iVar1 >= 0; iVar1 -= 10){
+    SetDungeonBGColorRGB(iVar1,0,0,1,1);
+    sub_803E46C(0x46);
+  }
+  sub_803E708(10,0x46);
+  sub_8085EB0();
+}
+
+void sub_808C550(void)
+{
+  gUnknown_202EDC8 = 0x1f;
+  SetDungeonBGColorRGB(0,0,0,1,1);
+  sub_803E9D0();
+  sub_803E46C(0x46);
+  gDungeonGlobalData->unk7 = 0;
+  DeoxysScreenFlash();
+}
+
+void sub_808C590(struct DungeonEntity *r0)
+{
+    sub_8068FE0(r0, 0x21C, 0);
+}
+
+void SetupDeoxysFacingDirection(struct DungeonEntity *deoxysEntity)
+{
+    SetFacingDirection(deoxysEntity, DIRECTION_SOUTH);
+}
+
+void SetupDeoxysFightHP(struct DungeonEntity *deoxysEntity)
+{
+    SetupBossFightHP(deoxysEntity, 950, 0x20);
+}
+
+void sub_808C5C0(void)
+{
+  struct DungeonEntity * LeaderEntity;
+  struct DungeonEntity * CelebiEntity;
+
+  u32 uVar3;
+  s32 iVar2;
+
+  LeaderEntity = xxx_call_GetLeaderEntity();
+  CelebiEntity = GetEntityFromClientType(0x1e);
+  sub_8083E88(0x7f);
+  sub_80854D4();
+  sub_8085930(DIRECTION_NORTH);
+  SetFacingDirection(CelebiEntity, DIRECTION_SOUTH);
+  sub_8085860(LeaderEntity->posWorld.x,LeaderEntity->posWorld.y);
+  uVar3 = GetCameraXPos();
+  iVar2 = GetCameraYPos();
+  sub_803F878(uVar3,iVar2 + 0xfffff000);
+  CopySpeciesNametoBuffer(gUnknown_202E038, SPECIES_CELEBI);
+}
+
+// Dummy Celebi Fight Dialogue??
+void nullsub_100(u32 r0, u32 r1, u32 r2)
+{
+}
+
+#ifdef NONMATCHING
+void CelebiJoinDialogue(void)
+{
+  struct DungeonEntity *LeaderEntity;
+  s32 state;
+  s32 menuChoice;
+  struct DungeonEntity *CelebiEntity;
+
+  LeaderEntity = xxx_call_GetLeaderEntity();
+  CelebiEntity = GetEntityFromClientType(0x1e);
+  if ((HasRecruitedMon(SPECIES_CELEBI)) || (sub_806FD18(CelebiEntity) == '\0'))
+  {
+    sub_8068FE0(local_1c,0x21c,0);
+    SpriteLookAroundEffect(LeaderEntity);
+    sub_803E708(10,0x46);
+    // .........
+    DisplayDungeonDialogue(&gCelebiJoinDialogue_10);
+  }
+  else
+  {
+    SpriteLookAroundEffect(LeaderEntity);
+    sub_803E708(10,0x46);
+    SpriteShockEffect(LeaderEntity);
+    sub_803E708(10,0x46);
+    // Oh? There's someone there.
+    DisplayDungeonDialogue(&gCelebiJoinDialogue_1);
+    sub_803E708(10,0x46);
+    sub_8086598();
+    sub_8086598();
+    sub_8086598();
+    sub_8086598();
+    // The Time-Traveling Pokemon {ARG_POKEMON_2} (Celebi)!
+    DisplayDungeonDialogue(&gCelebiJoinDialogue_2);
+    PlaySoundEffect(0x1c7);
+    sub_806CDD4(CelebiEntity,10,DIRECTION_SOUTH);
+    sub_803E708(0x14,0x46);
+    sub_806CE68(CelebiEntity, DIRECTION_SOUTH);
+    sub_803E708(4,0x46);
+    PlaySoundEffect(0x1c7);
+    sub_806CDD4(CelebiEntity,10,DIRECTION_SOUTH);
+    DisplayDungeonDialogue(&gCelebiJoinDialogue_3);
+    sub_803E708(10,0x46);
+    state = 0;
+    do {
+      switch(state)
+      {
+       case 0:
+            do {
+                // NOTE: 0 and Load of ptr switch order in compiler but else matches
+                menuChoice = sub_8052C68(0,gPtrPurityForestAllowCelebiToJoinText,
+                                &gPurityForestAllowCelebiToJoinPrompt,0x701);
+            } while (menuChoice < 1);
+            sub_803E708(10,0x46);
+            if (menuChoice == 1)
+            {
+                sub_8083F14();
+                PlaySoundEffect(0xcc);
+                while (sub_8083E74(0xcc) != 0) {
+                    sub_803E46C(0x46);
+                }
+                sub_8083E88(0x7f);
+                PlaySoundEffect(0x1c7);
+                sub_80861D4(CelebiEntity,0xd,0);
+                sub_803E708(0x37,0x46);
+                PlaySoundEffect(0x1d5);
+                sub_803E708(0x1a,0x46);
+                PlaySoundEffect(0x1d5);
+                sub_803E708(0x1c,0x46);
+                DisplayDungeonDialogue(&gCelebiJoinDialogue_4);
+                sub_803E708(10,0x46);
+                sub_806FDF4(LeaderEntity,CelebiEntity,&CelebiEntity);
+                sub_8083E88(0x7f);
+                DisplayDungeonDialogue(&gCelebiJoinDialogue_5);
+                sub_803E708(10,0x46);
+                PlaySoundEffect(0x1c7);
+                sub_80861D4(CelebiEntity,0xd,0);
+                sub_803E708(0x37,0x46);
+                PlaySoundEffect(0x1d5);
+                sub_803E708(0x1a,0x46);
+                PlaySoundEffect(0x1d5);
+                sub_803E708(0x1b,0x46);
+                DisplayDungeonDialogue(&gCelebiJoinDialogue_6);
+                sub_803E708(10,0x46);
+                state = 2;
+            }
+            else
+            {
+                state = 1;
+            }
+            break;
+        case 1:
+            do {
+                // NOTE: 0 and Load of ptr switch order in compiler but else matches
+                menuChoice = sub_8052C68(0,gPtrPurityForestRefuseCelebiConfirmText,
+                                    &gPurityForestRefuseCelebiConfirmPrompt,0x701);
+            } while (menuChoice < 1);
+            if (menuChoice == 1)
+                state = 0;
+            else
+            {
+                sub_803E708(10,0x46);
+                DisplayDungeonDialogue(&gCelebiJoinDialogue_7);
+                sub_80861F8(0x3e,CelebiEntity,1);
+                sub_803E708(0x18,0x46);
+                sub_80861F8(0x3e,CelebiEntity,1);
+                sub_803E708(0xe,0x46);
+                DisplayDungeonDialogue(&gCelebiJoinDialogue_8);
+                sub_803E708(10,0x46);
+                sub_808C8E0(CelebiEntity);
+                DisplayDungeonDialogue(&gCelebiJoinDialogue_9);
+                sub_803E708(10,0x46);
+                state = 2;
+            }
+            break;
+      }
+    } while (state != 2);
+  }
+  sub_8083ED8(0x1e);
+  sub_803E708(0x1e,0x46);
+  gDungeonGlobalData->unk2 = 1;
+}
+#else
+NAKED
+void CelebiJoinDialogue(void)
+{
+        asm_unified(
+        "\tpush {r4-r7,lr}\n"
+	"\tmov r7, r8\n"
+	"\tpush {r7}\n"
+	"\tsub sp, 0x4\n"
+	"\tbl xxx_call_GetLeaderEntity\n"
+	"\tadds r7, r0, 0\n"
+	"\tmovs r0, 0x1E\n"
+	"\tbl GetEntityFromClientType\n"
+	"\tstr r0, [sp]\n"
+	"\tmovs r0, 0x8A\n"
+	"\tlsls r0, 1\n"
+	"\tbl HasRecruitedMon\n"
+	"\tlsls r0, 24\n"
+	"\tcmp r0, 0\n"
+	"\tbne _0808C658\n"
+	"\tldr r0, [sp]\n"
+	"\tbl sub_806FD18\n"
+	"\tlsls r0, 24\n"
+	"\tcmp r0, 0\n"
+	"\tbne _0808C680\n"
+"_0808C658:\n"
+	"\tmovs r1, 0x87\n"
+	"\tlsls r1, 2\n"
+	"\tldr r0, [sp]\n"
+	"\tmovs r2, 0\n"
+	"\tbl sub_8068FE0\n"
+	"\tadds r0, r7, 0\n"
+	"\tbl SpriteLookAroundEffect\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r0, _0808C67C\n"
+	"\tbl DisplayDungeonDialogue\n"
+	"\tb _0808C8AC\n"
+	"\t.align 2, 0\n"
+"_0808C67C: .4byte gCelebiJoinDialogue_10\n"
+"_0808C680:\n"
+	"\tadds r0, r7, 0\n"
+	"\tbl SpriteLookAroundEffect\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tadds r0, r7, 0\n"
+	"\tbl SpriteShockEffect\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r0, _0808C71C\n"
+	"\tbl DisplayDungeonDialogue\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tbl sub_8086598\n"
+	"\tbl sub_8086598\n"
+	"\tbl sub_8086598\n"
+	"\tbl sub_8086598\n"
+	"\tldr r0, _0808C720\n"
+	"\tbl DisplayDungeonDialogue\n"
+	"\tldr r4, _0808C724\n"
+	"\tadds r0, r4, 0\n"
+	"\tbl PlaySoundEffect\n"
+	"\tldr r0, [sp]\n"
+	"\tmovs r1, 0xA\n"
+	"\tmovs r2, 0\n"
+	"\tbl sub_806CDD4\n"
+	"\tmovs r0, 0x14\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r0, [sp]\n"
+	"\tmovs r1, 0\n"
+	"\tbl sub_806CE68\n"
+	"\tmovs r0, 0x4\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tadds r0, r4, 0\n"
+	"\tbl PlaySoundEffect\n"
+	"\tldr r0, [sp]\n"
+	"\tmovs r1, 0xA\n"
+	"\tmovs r2, 0\n"
+	"\tbl sub_806CDD4\n"
+	"\tldr r0, _0808C728\n"
+	"\tbl DisplayDungeonDialogue\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tmovs r0, 0\n"
+	"\tmov r8, r4\n"
+	"\tldr r6, _0808C72C\n"
+"_0808C70E:\n"
+	"\tcmp r0, 0\n"
+	"\tbeq _0808C730\n"
+	"\tcmp r0, 0x1\n"
+	"\tbne _0808C718\n"
+	"\tb _0808C828\n"
+"_0808C718:\n"
+	"\tb _0808C8A6\n"
+	"\t.align 2, 0\n"
+"_0808C71C: .4byte gCelebiJoinDialogue_1\n"
+"_0808C720: .4byte gCelebiJoinDialogue_2\n"
+"_0808C724: .4byte 0x000001c7\n"
+"_0808C728: .4byte gCelebiJoinDialogue_3\n"
+"_0808C72C: .4byte 0x000001d5\n"
+"_0808C730:\n"
+	"\tldr r5, _0808C75C\n"
+"_0808C732:\n"
+	"\tmovs r0, 0\n"
+	"\tldr r1, [r5]\n"
+	"\tldr r2, _0808C760\n"
+	"\tldr r3, _0808C764\n"
+	"\tbl sub_8052C68\n"
+	"\tadds r4, r0, 0\n"
+	"\tcmp r4, 0\n"
+	"\tble _0808C732\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tcmp r4, 0x1\n"
+	"\tbne _0808C824\n"
+	"\tbl sub_8083F14\n"
+	"\tmovs r0, 0xCC\n"
+	"\tbl PlaySoundEffect\n"
+	"\tb _0808C76E\n"
+	"\t.align 2, 0\n"
+"_0808C75C: .4byte gPtrPurityForestAllowCelebiToJoinText\n"
+"_0808C760: .4byte gPurityForestAllowCelebiToJoinPrompt\n"
+"_0808C764: .4byte 0x00000701\n"
+"_0808C768:\n"
+	"\tmovs r0, 0x46\n"
+	"\tbl sub_803E46C\n"
+"_0808C76E:\n"
+	"\tmovs r0, 0xCC\n"
+	"\tbl sub_8083E74\n"
+	"\tlsls r0, 24\n"
+	"\tcmp r0, 0\n"
+	"\tbne _0808C768\n"
+	"\tmovs r0, 0x7F\n"
+	"\tbl sub_8083E88\n"
+	"\tmov r0, r8\n"
+	"\tbl PlaySoundEffect\n"
+	"\tldr r0, [sp]\n"
+	"\tmovs r1, 0xD\n"
+	"\tmovs r2, 0\n"
+	"\tbl sub_80861D4\n"
+	"\tmovs r0, 0x37\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tadds r0, r6, 0\n"
+	"\tbl PlaySoundEffect\n"
+	"\tmovs r0, 0x1A\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tadds r0, r6, 0\n"
+	"\tbl PlaySoundEffect\n"
+	"\tmovs r0, 0x1C\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r0, _0808C818\n"
+	"\tbl DisplayDungeonDialogue\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r1, [sp]\n"
+	"\tadds r0, r7, 0\n"
+	"\tmov r2, sp\n"
+	"\tbl sub_806FDF4\n"
+	"\tmovs r0, 0x7F\n"
+	"\tbl sub_8083E88\n"
+	"\tldr r0, _0808C81C\n"
+	"\tbl DisplayDungeonDialogue\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tmov r0, r8\n"
+	"\tbl PlaySoundEffect\n"
+	"\tldr r0, [sp]\n"
+	"\tmovs r1, 0xD\n"
+	"\tmovs r2, 0\n"
+	"\tbl sub_80861D4\n"
+	"\tmovs r0, 0x37\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tadds r0, r6, 0\n"
+	"\tbl PlaySoundEffect\n"
+	"\tmovs r0, 0x1A\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tadds r0, r6, 0\n"
+	"\tbl PlaySoundEffect\n"
+	"\tmovs r0, 0x1B\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r0, _0808C820\n"
+	"\tb _0808C898\n"
+	"\t.align 2, 0\n"
+"_0808C818: .4byte gCelebiJoinDialogue_4\n"
+"_0808C81C: .4byte gCelebiJoinDialogue_5\n"
+"_0808C820: .4byte gCelebiJoinDialogue_6\n"
+"_0808C824:\n"
+	"movs r0, 0x1\n"
+	"b _0808C8A6\n"
+"_0808C828:\n"
+	"\tldr r5, _0808C844\n"
+"_0808C82A:\n"
+	"\tmovs r0, 0\n"
+	"\tldr r1, [r5]\n"
+	"\tldr r2, _0808C848\n"
+	"\tldr r3, _0808C84C\n"
+	"\tbl sub_8052C68\n"
+	"\tadds r4, r0, 0\n"
+	"\tcmp r4, 0\n"
+	"\tble _0808C82A\n"
+	"\tcmp r4, 0x1\n"
+	"\tbne _0808C850\n"
+	"\tmovs r0, 0\n"
+	"\tb _0808C8A6\n"
+	"\t.align 2, 0\n"
+"_0808C844: .4byte gPtrPurityForestRefuseCelebiConfirmText\n"
+"_0808C848: .4byte gPurityForestRefuseCelebiConfirmPrompt\n"
+"_0808C84C: .4byte 0x00000701\n"
+"_0808C850:\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r0, _0808C8D0\n"
+	"\tbl DisplayDungeonDialogue\n"
+	"\tldr r1, [sp]\n"
+	"\tmovs r0, 0x3E\n"
+	"\tmovs r2, 0x1\n"
+	"\tbl sub_80861F8\n"
+	"\tmovs r0, 0x18\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r1, [sp]\n"
+	"\tmovs r0, 0x3E\n"
+	"\tmovs r2, 0x1\n"
+	"\tbl sub_80861F8\n"
+	"\tmovs r0, 0xE\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r0, _0808C8D4\n"
+	"\tbl DisplayDungeonDialogue\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r0, [sp]\n"
+	"\tbl sub_808C8E0\n"
+	"\tldr r0, _0808C8D8\n"
+"_0808C898:\n"
+	"\tbl DisplayDungeonDialogue\n"
+	"\tmovs r0, 0xA\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tmovs r0, 0x2\n"
+"_0808C8A6:\n"
+	"\tcmp r0, 0x2\n"
+	"\tbeq _0808C8AC\n"
+	"\tb _0808C70E\n"
+"_0808C8AC:\n"
+	"\tmovs r0, 0x1E\n"
+	"\tbl sub_8083ED8\n"
+	"\tmovs r0, 0x1E\n"
+	"\tmovs r1, 0x46\n"
+	"\tbl sub_803E708\n"
+	"\tldr r0, _0808C8DC\n"
+	"\tldr r1, [r0]\n"
+	"\tmovs r0, 0x1\n"
+	"\tstrb r0, [r1, 0x2]\n"
+	"\tadd sp, 0x4\n"
+	"\tpop {r3}\n"
+	"\tmov r8, r3\n"
+	"\tpop {r4-r7}\n"
+	"\tpop {r0}\n"
+	"\tbx r0\n"
+	"\t.align 2, 0\n"
+"_0808C8D0: .4byte gCelebiJoinDialogue_7\n"
+"_0808C8D4: .4byte gCelebiJoinDialogue_8\n"
+"_0808C8D8: .4byte gCelebiJoinDialogue_9\n"
+"_0808C8DC: .4byte gDungeonGlobalData");
+}
+#endif
+
+void sub_808C8E0(struct DungeonEntity *param_1)
+{
+  s32 iVar1;
+
+  PlaySoundEffect(0x1a5);
+  sub_806CDD4(param_1, 0, DIRECTION_SOUTH);
+  for(iVar1 = 0; iVar1 < 16; iVar1++){
+    param_1->entityData->unk174 = iVar1 * 256;
+    sub_803E46C(0x46);
+  }
+  for(iVar1 = 16; iVar1 < 200; iVar1 += 4){
+    param_1->entityData->unk174 = iVar1 * 256;
+    sub_803E46C(0x46);
+  }
+  sub_8086A3C(param_1);
+}
+
+void sub_808C938(void)
+{
+    sub_80855E4(sub_808C9B0);
+}
+
+void sub_808C948(struct DungeonEntity *param_1, u8 param_2)
+{
+  bool8 flag;
+  s32 iVar3;
+  struct DungeonEntity *iVar2;
+
+  if (param_2 == 0x37) {
+    flag = FALSE;
+    for(iVar3 = 0; iVar3 < DUNGEON_MAX_WILD_POKEMON; iVar3++){
+      iVar2 = gDungeonGlobalData->wildPokemon[iVar3];
+      if ((iVar2 != param_1) && (EntityExists(iVar2) != 0)) {
+        flag = TRUE;
+        break;
+      }
+    }
+    if (!flag) {
+      gDungeonGlobalData->unk2 = 1;
+    }
+  }
+}
+
+void sub_808C998(void)
+{
+    // Defeat the opposing team to win
+    // Be careful, your opponents are tough
+    DisplayDungeonDialogue(&gUnknown_8106720);
+    sub_803E708(0xA, 0x46);
+}
+
+void sub_808C9B0(struct DungeonEntity *param_1)
+{
+    param_1->entityData->action.facingDir = DIRECTION_NORTH;
+    sub_806CE68(param_1, DIRECTION_NORTH);
+}
+
+void sub_808C9C4(void)
+{
+    struct DungeonEntity *LeaderEntity;
+    struct DungeonEntity *MedichamEntity;
+
+    LeaderEntity = xxx_call_GetLeaderEntity();
+    MedichamEntity = GetEntityFromClientType(7);
+    sub_8083E88(0x72);
+    sub_80854D4();
+    sub_8085930(DIRECTION_NORTH);
+    sub_80855E4(sub_8086A3C);
+    SetFacingDirection(MedichamEntity, DIRECTION_SOUTH);
+    sub_8085860(LeaderEntity->posWorld.x, LeaderEntity->posWorld.y - 3);
+    CopySpeciesNametoBuffer(gUnknown_202E038, SPECIES_MEDICHAM);
+}
+
+// Medicham Rescue Dialogue?
+void MedichamRescueDialogue(void)
+{
+    struct DungeonEntity *MedichamEntity;
+    s32 counter;
+
+    MedichamEntity = GetEntityFromClientType(7);
+    SpriteLookAroundEffect(MedichamEntity);
+    sub_803E708(0xA, 0x46);
+    // Oh my I can't seem to find a way out...
+    DisplayDungeonDialogue(&gMedichamRescueDialogue_1);
+    sub_803E708(0xA, 0x46);
+    sub_80869E4(MedichamEntity, 4, 2, 4);
+    sub_803E708(0xA, 0x46);
+    // What am I to do...?
+    DisplayDungeonDialogue(&gMedichamRescueDialogue_2);
+    sub_803E708(0xA, 0x46);
+    sub_8086448();
+    sub_8086598();
+    sub_803E708(0x20, 0x46);
+    SpriteShockEffect(MedichamEntity);
+    sub_803E708(0x20, 0x46);
+    sub_80869E4(MedichamEntity, 4, 2, 0);
+    DisplayDungeonDialogue(&gMedichamRescueDialogue_3);
+    sub_803E708(0xA, 0x46);
+    PlaySoundEffect(0x1c7);
+    sub_806CDD4(MedichamEntity, 0xA, DIRECTION_SOUTH);
+    sub_803E708(0x14, 0x46);
+    sub_806CE68(MedichamEntity, DIRECTION_SOUTH);
+    sub_803E708(0x4, 0x46);
+    PlaySoundEffect(0x1c7);
+    sub_806CDD4(MedichamEntity, 0xA, DIRECTION_SOUTH);
+    sub_803E708(0x14, 0x46);
+    // Yes Yes
+    // I am so lucky
+    //
+    // There appears to be no one here
+    DisplayDungeonDialogue(&gMedichamRescueDialogue_4);
+    sub_803E708(0xA, 0x46);
+    sub_806CDD4(MedichamEntity, 0, DIRECTION_SOUTH);
+    for(counter = 0x17; counter >= 0; counter--)
+    {
+        sub_804539C(MedichamEntity, 0, 0x80 << 1);
+        sub_803E46C(0x46);
+    }
+    sub_806CE68(MedichamEntity, DIRECTION_SOUTH);
+    sub_803E708(0x20, 0x46);
+    sub_8042B0C(MedichamEntity);
+    sub_8068FE0(MedichamEntity, 0x21C, 0);
+    gDungeonGlobalData->unk4 = 1;
+    gDungeonGlobalData->unk11 = 4;
+}
+
+void sub_808CB5C(void)
+{
+    struct DungeonEntity *LeaderEntity;
+    struct DungeonEntity *MedichamEntity;
+
+    LeaderEntity = xxx_call_GetLeaderEntity();
+    MedichamEntity = GetEntityFromClientType(7);
+    sub_8083E88(0x72);
+    sub_80854D4();
+    sub_8085930(DIRECTION_NORTH);
+    sub_8068FE0(MedichamEntity, 0x21C, 0);
+    sub_8085860(LeaderEntity->posWorld.x, LeaderEntity->posWorld.y);
+    CopySpeciesNametoBuffer(gUnknown_202E038, SPECIES_MEDICHAM);
+}
+
+void DummyFightDialogue(void)
+{
+}
+
+void sub_808CBB0(void)
+{
+    struct DungeonEntity *LeaderEntity;
+    struct DungeonEntity *SmeargleEntity;
+
+    LeaderEntity = xxx_call_GetLeaderEntity();
+    SmeargleEntity = GetEntityFromClientType(0x1F);
+    sub_8083E88(0x72);
+    sub_80854D4();
+    sub_8085930(DIRECTION_NORTH);
+    sub_80855E4(sub_8086A3C);
+    SetFacingDirection(SmeargleEntity, DIRECTION_SOUTH);
+    sub_8085860(LeaderEntity->posWorld.x, LeaderEntity->posWorld.y - 3);
+    CopySpeciesNametoBuffer(gUnknown_202E038, SPECIES_SMEARGLE);
+}
+
+// Smeargle Rescue dialogue scene
+void SmeargleRescueDialogue(void)
+{
+    struct DungeonEntity *SmeargleEntity;
+    s32 counter;
+
+    SmeargleEntity = GetEntityFromClientType(0x1F);
+    SpriteLookAroundEffect(SmeargleEntity);
+    sub_803E708(0xA, 0x46);
+    // Ohhh...
+    // I've lost my bearings
+    DisplayDungeonDialogue(&gSmeargleRescueDialogue_1);
+    sub_803E708(0xA, 0x46);
+    sub_80869E4(SmeargleEntity, 4, 2, 4);
+    sub_803E708(0xA, 0x46);
+    // I can't get out...
+    // I'm hungry...
+    // I'm in trouble
+    DisplayDungeonDialogue(&gSmeargleRescueDialogue_2);
+    sub_803E708(0xA, 0x46);
+    sub_8086448();
+    sub_8086598();
+    sub_803E708(0x20, 0x46);
+    SpriteShockEffect(SmeargleEntity);
+    sub_803E708(0x20, 0x46);
+    sub_80869E4(SmeargleEntity, 4, 2, 0);
+    // Oh! You are?
+    DisplayDungeonDialogue(&gSmeargleRescueDialogue_3);
+    sub_803E708(0xA, 0x46);
+    PlaySoundEffect(0x1c7);
+    sub_806CDD4(SmeargleEntity, 0xA, DIRECTION_SOUTH);
+    sub_803E708(0x14, 0x46);
+    sub_806CE68(SmeargleEntity, DIRECTION_SOUTH);
+    sub_803E708(0x4, 0x46);
+    PlaySoundEffect(0x1c7);
+    sub_806CDD4(SmeargleEntity, 0xA, DIRECTION_SOUTH);
+    sub_803E708(0x14, 0x46);
+    // Did you maybe come to rescue me?
+    // Am I glad to see you
+    DisplayDungeonDialogue(&gSmeargleRescueDialogue_4);
+    sub_803E708(0xA, 0x46);
+    sub_806CDD4(SmeargleEntity, 0, DIRECTION_SOUTH);
+    for(counter = 0x17; counter >= 0; counter--)
+    {
+        sub_804539C(SmeargleEntity, 0, 0x80 << 1);
+        sub_803E46C(0x46);
+    }
+    sub_806CE68(SmeargleEntity, DIRECTION_SOUTH);
+    sub_803E708(0x20, 0x46);
+    sub_8042B0C(SmeargleEntity);
+    sub_8068FE0(SmeargleEntity, 0x21C, 0);
+    gDungeonGlobalData->unk4 = 1;
+    gDungeonGlobalData->unk11 = 4;
+
+}
+
+void sub_808CD44(void)
+{
+    struct DungeonEntity *LeaderEntity;
+    struct DungeonEntity *SmeargleEntity;
+
+    LeaderEntity = xxx_call_GetLeaderEntity();
+    SmeargleEntity = GetEntityFromClientType(0x1F);
+    sub_8083E88(0x72);
+    sub_80854D4();
+    sub_8085930(DIRECTION_NORTH);
+    sub_80855E4(sub_8086A3C);
+    sub_8068FE0(SmeargleEntity, 0x21C, 0);
+    sub_8085860(LeaderEntity->posWorld.x, LeaderEntity->posWorld.y - 3);
+    CopySpeciesNametoBuffer(gUnknown_202E038, SPECIES_SMEARGLE);
+}
+
+void sub_808CD9C(u8 r0)
+{
+    sub_8086448();
+    // There appears to be no one here.
+    // It's impossible to go any further
+    sub_80866C4(&gUnknown_810697C);
+}
+
+u64 *sub_808CDB0(u8 r0)
+{
+    return &gUnknown_8107544[r0];
 }
